feat: cleaning

+++ relayer/src/services/identityService.ts
@@ -7,7 +7,7 @@ import EnsService from './ensService.js';
 import AuthorisationService from './authorisationService.js';
 
 const defaultDeployOptions = {
-  gasLimit: utils.bigNumberify(3500000),
+  gasLimit: utils.bigNumberify(5500000),
   gasPrice: utils.bigNumberify(9000000000)
 };
 
@@ -24,7 +24,7 @@ export default class IdentityService {
     this.petitionToken = jsonRpcService.contracts.ERC20Mintable[0];
   }
 
-  async create(address: string, ensName: string, overrideOptions = {}) {
+  async create(address: string, ensName: string) {
     const key = addressToBytes32(address);
     const ensArgs = await this.ensService.argsFor(ensName);
     if (!ensArgs) {
@@ -34,7 +34,6 @@ export default class IdentityService {
     const args = [key, ...ensArgs];
     const deployTransaction = {
       ...defaultDeployOptions,
-      ...overrideOptions,
       ...new ContractFactory(this.abi, this.bytecode).getDeployTransaction(...args)
     };
     return await this.jsonRpcService.wallet.sendTransaction(deployTransaction);
@@ -47,54 +46,17 @@ export default class IdentityService {
 
     const transaction = this.buildTransaction(message);
     const { wallet, provider } = this.jsonRpcService;
-    // const estimateGas = await provider.estimateGas({...transaction, from: wallet.address});
-    // if (!utils.bigNumberify(message.gasLimit).gte(estimateGas)) {
-    //   throw new Error('Not enough Gas');
-    // }
+    const estimateGas = await provider.estimateGas({...transaction, from: wallet.address});
+    if (!utils.bigNumberify(message.gasLimit).gte(estimateGas)) {
+      throw new Error('Not enough Gas');
+    }
 
     if (message.to === message.from && this.isAddKeyCall(message.data)) {
       const key = this.getKeyFromData(message.data);
       this.authorisationService.removeRequest(message.from, key);
-      return await wallet.sendTransaction(transaction);
-    } else if (message.to === message.from && this.isAddKeysCall(message.data)) {
-      return await wallet.sendTransaction(transaction);
     }
-
-    const contract = new ethers.Contract(message.from, Identity.abi, wallet);
-    console.dir("MESSAGE")
-    console.dir([
-      message.to,
-      0,
-      message.data,
-      0,
-      1000000000,
-      message.gasToken,
-      1000000,
-      0,
-      message.signature
-    ])
-    // address to,
-    //     uint256 value,
-    //     bytes memory data,
-    //     uint nonce,
-    //     uint gasPrice,
-    //     address gasToken,
-    //     uint gasLimit,
-    //     OperationType operationType,
-    //     bytes memory signatures
-    return await contract.executeSigned(
-      message.to,
-      0,
-      message.data,
-      0,
-      1000000000,
-      message.gasToken,
-      1000000,
-      0,
-      message.signature
-    )
     
-    // return await this.jsonRpcService.wallet.sendTransaction(transaction);
+    return await wallet.sendTransaction(transaction);
   }
 
   private buildTransaction(message: any) {
@@ -103,9 +65,9 @@ export default class IdentityService {
       message.value,
       message.data,
       message.nonce,
-      message.gasPrice,
+      utils.bigNumberify(message.gasPrice),
       message.gasToken,
-      message.gasLimit,
+      utils.bigNumberify(message.gasLimit),
       message.operationType,
       message.signature
     ]);
@@ -121,11 +83,6 @@ export default class IdentityService {
   private isAddKeyCall(data: string) {
     const addKeySighash = new utils.Interface(Identity.abi).functions.addKey.sighash;
     return addKeySighash === data.slice(0, addKeySighash.length);
-  };
-  
-  private isAddKeysCall(data: string) {
-    const addKeysSighash = new utils.Interface(Identity.abi).functions.addKeys.sighash;
-    return addKeysSighash === data.slice(0, addKeysSighash.length);
   };
 
   private getKeyFromData(data: string) {
